+++ packages/electron-better-sqlite3-backend/src/electronBetterSqlite3Backend.ts
@@ -1,4 +1,4 @@
-import { IDbBackend, IQuery, IQueryResult } from "@trong-orm/core";
+import { IDbBackend, IQuery, IQueryResult } from "@kikko-land/core";
 
 declare global {
   interface Window {