+++ packages/electron-better-sqlite3-backend/vite.config.js
@@ -1,4 +1,4 @@
-const { buildConfig } = require("@trong-orm/common-scripts/vite.cjs");
+const { buildConfig } = require("@kikko-land/common-scripts/vite.cjs");
 
 module.exports = buildConfig({
   ...(() => {