+++ packages/query-builder/src/tokens/statements/select.ts
@@ -1,4 +1,4 @@
-import { ISql, ISqlAdapter, sql } from "@trong-orm/sql";
+import { ISql, ISqlAdapter, sql } from "@kikko-land/sql";
 
 import { IBaseToken, isToken, TokenType } from "../../types";
 import { alias } from "../alias";