+++ packages/react-native-backend/src/reactNativeBackend.ts
@@ -1,4 +1,4 @@
-import { IDbBackend, IQuery, IQueryResult } from "@trong-orm/core";
+import { IDbBackend, IQuery, IQueryResult } from "@kikko-land/core";
 import {
   Location,
   openDatabase,