+++ packages/vite-react-example/src/hooks/usePaginator.ts
@@ -1,5 +1,5 @@
-import { ISelectStatement, select } from "@trong-orm/query-builder";
-import { sql, useQueryFirstRow } from "@trong-orm/react";
+import { ISelectStatement, select } from "@kikko-land/query-builder";
+import { sql, useQueryFirstRow } from "@kikko-land/react";
 import { useCallback, useEffect, useState } from "react";
 
 export const usePaginator = ({